refactor: rearrange directory structure and add Controller/Service interface
