+++ src/main/java/nextstep/ladder/application/LadderGameService.java
@@ -1,0 +1,18 @@
+package nextstep.ladder.application;
+
+import nextstep.ladder.application.view.InputView;
+import nextstep.ladder.domain.ladder.Ladder;
+import nextstep.ladder.util.painter.PaintingStrategy;
+
+public class LadderGameService {
+
+	private final PaintingStrategy paintingStrategy;
+
+	public LadderGameService(PaintingStrategy paintingStrategy) {
+		this.paintingStrategy = paintingStrategy;
+	}
+
+	public Ladder getLadder() {
+		return Ladder.ofLadder(InputView.getPlayers(), InputView.getHeights(), paintingStrategy);
+	}
+}
