--- src/main/java/nextstep/ladder/ladder/Ladder.java
@@ -1,45 +1,0 @@
-package nextstep.ladder.ladder;
-
-import static java.util.stream.Collectors.*;
-
-import java.util.stream.IntStream;
-
-import nextstep.ladder.height.Height;
-import nextstep.ladder.line.Line;
-import nextstep.ladder.line.Lines;
-import nextstep.ladder.painter.PaintingStrategy;
-import nextstep.ladder.player.Players;
-
-public class Ladder {
-
-	private final Players players;
-	private final Lines lines;
-	private final PaintingStrategy paintingStrategy;
-
-	private Ladder(Players players, Height height, PaintingStrategy paintingStrategy) {
-		this.players = players;
-		this.paintingStrategy = paintingStrategy;
-		this.lines = drawLines(height);
-	}
-
-	public static Ladder ofLadder(Players players, Height height, PaintingStrategy paintingStrategy) {
-		return new Ladder(players, height, paintingStrategy);
-	}
-
-	private Lines drawLines(Height height) {
-		int playerCount = players.getSize();
-		int heightCount = height.getHeightValue();
-
-		return IntStream.range(0, heightCount)
-			.mapToObj(index -> Line.ofLine(playerCount, paintingStrategy))
-			.collect(collectingAndThen(toList(), Lines::ofLines));
-	}
-
-	public Players getPlayers() {
-		return players;
-	}
-
-	public Lines getLines() {
-		return lines;
-	}
-}
