--- src/main/java/nextstep/ladder/painter/LinePainter.java
@@ -1,10 +1,0 @@
-package nextstep.ladder.painter;
-
-public class LinePainter {
-
-	private final PointPainter pointPainter;
-
-	public LinePainter(PointPainter pointPainter) {
-		this.pointPainter = pointPainter;
-	}
-}
