--- src/main/java/nextstep/ladder/painter/PaintingStrategy.java
@@ -1,7 +1,0 @@
-package nextstep.ladder.painter;
-
-import nextstep.ladder.point.Point;
-
-public interface PaintingStrategy {
-	Point drawPoint();
-}
