--- src/main/java/nextstep/ladder/painter/RandomPaintingStrategy.java
@@ -1,15 +1,0 @@
-package nextstep.ladder.painter;
-
-import java.util.Random;
-
-import nextstep.ladder.point.Point;
-
-public class RandomPaintingStrategy implements PaintingStrategy {
-
-	private final Random random = new Random();
-
-	@Override
-	public Point drawPoint() {
-		return Point.ofPoint(random.nextBoolean());
-	}
-}
