+++ src/main/java/nextstep/ladder/util/painter/LinePainter.java
@@ -1,0 +1,10 @@
+package nextstep.ladder.util.painter;
+
+public class LinePainter {
+
+	private final PointPainter pointPainter;
+
+	public LinePainter(PointPainter pointPainter) {
+		this.pointPainter = pointPainter;
+	}
+}
