+++ src/main/java/nextstep/ladder/util/painter/PaintingStrategy.java
@@ -1,0 +1,7 @@
+package nextstep.ladder.util.painter;
+
+import nextstep.ladder.domain.point.Point;
+
+public interface PaintingStrategy {
+	Point drawPoint();
+}
